Major look and feel updates underway
This commit has a bunch of extra stuff in the comments of Layout as well as unused components which sort of map the progression of developing what is more or less the final version of a perlin-esque noise field generating semi-randomly displayed foreground vs background pixels on a threejs canvas. Developing this graphic took several days, because I am not versed in threejs or shader math. I had the idea of what I wanted based on something I saw somewhere else, but went through several interesting iterations trying to recreate. The process of doing this was interesting enough to me to document in an entire commit. The following commit will be mostly a refactor with minor touchups as I decide how to make the rest of the site fit visually with this perlin noise particle cloud effect.

This commit also has the disorganized learning process for implementing my sliding graphic effect. The long and short of that process was, making the graphic slide scroll-proportional, then making the graphic stay put once you scroll past, then making the graphic slide back out once you scroll back up, then finally making sure that the slide wasn't incomplete if you scrolled past too quickly. The files don't reflect this process, and I should have been committing throughout, but the files do reflect an ambivalence around whether to use react for the effect or to try and do it in base astro. Astro is surprisingly annoying about doing visual effects with js/ts, so it made more sense (and upset the ts compiler less) when I just sucked it up and made it a react component. (This leniency also paved the way for me to acquiesce to the use of threejs in the perlin noise effect as well.)